Attempt to include special code only available for cli_examples
The cli_examples are converted also to unit tests. These unit tests will
compile a binary and execute it. The main goal is to enforce
documentation correctness through the unit tests.

Sometimes a special case should be hard-coded (especially mocking user
input) that should only be available for the binaries to be executed
from the unit tests but should be excluded otherwise.

This PR adds the ability to include code through option_env! macro since
the CLI_TEST should only be defined in for the binary to be executed
from the unit tests.

This PR is a pre-requisite to solve https://github.com/FuelLabs/forc-wallet/issues/157

This PR will include more examples as part of #5444